+++ DownFall/Extensions/SKNode+Extension.swift
@@ -12,7 +12,5 @@ extension SKNode {
     func wasTouched(_ touches: Set<UITouch>, with event: UIEvent?) -> Bool {
         guard let _ = touches.first else { return false }
         return true
-//        let nodes = self.nodes(at: touch.location(in: self))
-//        return nodes.contains { $0.name == self.name }
     }
 }
